REGRESSION (r219342): Touch event coordinates and elementFromPoint coordinates differ https://bugs.webkit.org/show_bug.cgi?id=182910 <rdar://problem/37533950>
Reviewed by Simon Fraser.

Source/WebCore:

We reverted other changes to the definition of client coordinates
in r219829 due to compatibility concerns. However, we failed to revert
r219342 on trunk, leaving elementFromPoint() using coordinates relative
to the layout viewport.

Add a currently off-by-default setting to switch on layout-viewport-relative
client coordinates and guard the elementFromPoint changes behind it.
A future patch should roll r219829 back in also behind this setting, so
that everything remains consistent regardless of which coordinate space we choose.

* dom/TreeScope.cpp:
(WebCore::absolutePointIfNotClipped):
* page/Settings.yaml:
* rendering/RenderLayer.cpp:
(WebCore::RenderLayer::hitTest):

LayoutTests:

* fast/dom/elementFromPoint-scaled-scrolled.html:
Revert changes to this test made in r219342.

* fast/dom/elementFromPoint-scaled-scrolled-layout-viewport.html:
* fast/dom/elementFromPoint-scaled-scrolled-layout-viewport-expected.txt:
Add a test that is equivalent to elementFromPoint-scaled-scrolled.html after r219342,
which turns on the new setting. This test is disabled on iOS (like it was
in r219342) because it needs window.scrollTo.

* platform/ios-wk2/fast/dom/elementFromPoint-relative-to-viewport-expected.txt:
This now passes.

* platform/ios/TestExpectations:
Re-mark-failing a test that was un-marked-failing by r219342.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@228714 268f45cc-cd09-0410-ab3c-d52691b4dbfc