AX: Flaky test after r229310 https://bugs.webkit.org/show_bug.cgi?id=183376 <rdar://problem/38188685>
Reviewed by Chris Fleizach.

Source/WebCore:

The test that forces the context menu to show sometimes locks up
the test runner. Also, we missed a case where calling AXShowMenu
action on a combobox is not dispatching the accessibility event.

Using the combobox example to test the event dispatching in order to
avoid the test hang.

* accessibility/mac/WebAccessibilityObjectWrapperMac.mm:
(-[WebAccessibilityObjectWrapper accessibilityPerformShowMenuAction]):
(-[WebAccessibilityObjectWrapper accessibilityShowContextMenu]):

LayoutTests:

* accessibility/mac/AOM-events-all-expected.txt:
* accessibility/mac/AOM-events-all.html:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@229344 268f45cc-cd09-0410-ab3c-d52691b4dbfc