[TexMap] Clean up TextureMapperLayer constructor https://bugs.webkit.org/show_bug.cgi?id=183634
Reviewed by Michael Catanzaro.

In the TextureMapperLayer header, only declare the constructor, and move
all the previous member initialization to the point of declaration of
these member variables.

In the TextureMapperLayer implementation file, the constructor and
destructor are moved into proper order at the top of the file. The
former is defaulted, and the latter remains unchanged.

No change in behavior.

* platform/graphics/texmap/TextureMapperLayer.cpp:
(WebCore::TextureMapperLayer::~TextureMapperLayer):
* platform/graphics/texmap/TextureMapperLayer.h:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@229670 268f45cc-cd09-0410-ab3c-d52691b4dbfc