Correct debug assertion in Range::borderAndTextRects https://bugs.webkit.org/show_bug.cgi?id=183710 <rdar://problem/38466976>
Reviewed by Ryosuke Niwa.

Source/WebCore:

A debug assertion will fire if Range::borderAndTextRects is asked to evaluate a set of selected elements, where one of the elements does not have a parent.

We should consider a nullptr parent as satisfying the condition of a parent not being present in the selection set.

Tests: fast/dom/range/range-selection-empty-body.html

* dom/Range.cpp:
(WebCore::Range::borderAndTextRects const):

LayoutTests:

* fast/dom/Range/range-selection-empty-body-expected.txt: Added.
* fast/dom/Range/range-selection-empty-body.html: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@229695 268f45cc-cd09-0410-ab3c-d52691b4dbfc