[Cairo] Draw Cairo patterns with cairo_paint_with_alpha() https://bugs.webkit.org/show_bug.cgi?id=183774
Reviewed by Carlos Garcia Campos.

In the drawPatternToCairoContext() helper in CairoOperations.cpp source
file, always clip the painting region to the specified rectangle and
then use cairo_paint_with_alpha() to paint the passed-in pattern object,
not relaying rasterization to cairo_fill() when using opaque alpha. We
still clamp the alpha value to the 0.0 - 1.0 range.

No new tests -- no change in behavior.

* platform/graphics/cairo/CairoOperations.cpp:
(WebCore::Cairo::drawPatternToCairoContext):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@229796 268f45cc-cd09-0410-ab3c-d52691b4dbfc