[iOS] Text highlighted by the Find UI overlaps with NBC news header on google.com https://bugs.webkit.org/show_bug.cgi?id=183658
Patch by Frederic Wang <fwang@igalia.com> on 2018-03-21
Reviewed by Tim Horton.

Source/WebKit:

On Mobile, many sites have overlay headers or footers that may overlap with the highlighted
text inside a scrollable overflow node. To workaround that issue, this commit changes the
behavior so that the text is revealed at the center of the overflow node.

* WebProcess/WebPage/ios/FindControllerIOS.mm:
(WebKit::FindController::didFindString): Use alignCenterAlways instead of alignToEdgeIfNeeded.

LayoutTests:

Improve the test added in r224284, so that we test when a result is revealed at
the center of the overflow node.

* platform/ios/fast/scrolling/find-text-in-overflow-node.html: Put some content after
'match 4' so that the overflow node can scroll and make the text appear at the center.
Perform more accurate testing of the scrollTop of the overflow node.
* platform/ios/fast/scrolling/find-text-in-overflow-node-expected.txt: Update expectation.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@229801 268f45cc-cd09-0410-ab3c-d52691b4dbfc