Web Automation: clipToViewport is ignored for element screenshots https://bugs.webkit.org/show_bug.cgi?id=184158 <rdar://problem/39014307>
Reviewed by Timothy Hatcher.

In §19.2 Take Element Screenshot, step 5.2 says that we should clip
the element screenshot rect with the visible viewport rect. We don't
do that right now even though we pass over clipToViewport.

* WebProcess/Automation/WebAutomationSessionProxy.cpp:
(WebKit::snapshotRectForScreenshot):
Clip the rect to viewport if needed.

(WebKit::WebAutomationSessionProxy::takeScreenshot):
This scrollIntoView is misplaced; by this point we have already done
the math to figure out the screenshot rect. Move it before computing the rect.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@230088 268f45cc-cd09-0410-ab3c-d52691b4dbfc