totalPhysicalSize calculation when splitting a range must account for double rounding effects https://bugs.webkit.org/show_bug.cgi?id=184275
Reviewed by Mark Lam.

The rounding error could happen when we split a range where the
range's total physical size equals the range's total size. The
rounding may cause the left size to lose a byte, and the right
size to gain a byte. This caused the right side to be a byte
large than its size.

* bmalloc/LargeRange.h:
(bmalloc::LargeRange::LargeRange):
(bmalloc::LargeRange::split const):



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@230222 268f45cc-cd09-0410-ab3c-d52691b4dbfc