Don't have selections get stuck in editable elements https://bugs.webkit.org/show_bug.cgi?id=184483
Reviewed by Timothy Hatcher.

When selecting in editable elements, if the keyboard does not come up, which is not required,
then the selections are stuck in that element. If you try and select something in web content,
it will only select in the editable element. By clearing the assisted node when clearing a selection,
it no longer sticks. Also, textSelectionAssistant should not be used as if we are currently editing.
Change it to the function we created for that purpose in a pervious patch.

* UIProcess/ios/WKContentViewInteraction.mm:
(-[WKContentView clearSelection]):
(-[WKContentView _isInteractingWithAssistedNode]):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@230588 268f45cc-cd09-0410-ab3c-d52691b4dbfc