[macOS] Don't establish unneeded Dock connections https://bugs.webkit.org/show_bug.cgi?id=184664 <rdar://problem/16863698>
Reviewed by Simon Fraser.

There is no reason for the WebContent or Plugin processes to interact with
the Dock. We should tell AppKit that we don't want this connection, and to
avoid creating such connections.

Source/WebCore/PAL:

* pal/spi/mac/NSApplicationSPI.h: Add Dock connection declaration.

Source/WebKit:

* PluginProcess/mac/PluginProcessMac.mm:
(WebKit::PluginProcess::platformInitializeProcess): Tell NSApplication to
not create a Dock connection.
* WebProcess/cocoa/WebProcessCocoa.mm:
(WebKit::WebProcess::platformInitializeWebProcess): Ditto.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@230689 268f45cc-cd09-0410-ab3c-d52691b4dbfc