REGRESSION: [mac-wk2 release] LayoutTest http/tests/security/contentSecurityPolicy/script-src-blocked-error-event.html is flaky https://bugs.webkit.org/show_bug.cgi?id=184673 <rdar://problem/39474698>
Unreviewed test gardening. The test was flaky because the HTTP parser
would try to preload the Javascript resource. Sometimes this operation
would complete in time to print its own console message, duplicating a
subsequent one when the non-preloading resource-load occurs. Updated
the test to load the Javascript in such a way that it skips the reload
step.

* TestExpectations:
* http/tests/security/contentSecurityPolicy/script-src-blocked-error-event-expected.txt:
* http/tests/security/contentSecurityPolicy/script-src-blocked-error-event.html:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@230691 268f45cc-cd09-0410-ab3c-d52691b4dbfc