[GTK][WPE] TestSSL fails due to additional TLS errors returned https://bugs.webkit.org/show_bug.cgi?id=184860
Reviewed by Carlos Garcia Campos.

Add script and config file for regenerating the test certificate.
Regenerate it. TestSSL no longer fails on my machine.

I do see a ton of network process crashes and JavaScript errors, none of
which prevent the test from passing, but this commit only claims to fix
the certificate validation portion of the test.

* TestWebKitAPI/Tests/WebKitGLib/resources/generate-test-cert.sh: Added.
* TestWebKitAPI/Tests/WebKitGLib/resources/test-cert.conf: Added.
* TestWebKitAPI/Tests/WebKitGLib/resources/test-cert.pem:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@230911 268f45cc-cd09-0410-ab3c-d52691b4dbfc