feTurbulence is not rendered correctly on Retina display https://bugs.webkit.org/show_bug.cgi?id=183798
Patch by Said Abou-Hallawa <sabouhallawa@apple.com> on 2018-05-08
Reviewed by Simon Fraser.

Source/WebCore:

On 2x display the feTurbulence filter creates a scaled ImageBuffer but
processes only the unscaled size. This is a remaining work of r168577 and
is very similar to what was done for the feMorphology filter in r188271.

Test: fast/hidpi/filters-turbulence.html

* platform/graphics/filters/FETurbulence.cpp:
(WebCore::FETurbulence::fillRegion const):
(WebCore::FETurbulence::platformApplySoftware):

LayoutTests:

* fast/hidpi/filters-turbulence-expected.html: Added.
* fast/hidpi/filters-turbulence.html: Added.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@231485 268f45cc-cd09-0410-ab3c-d52691b4dbfc