REGRESSION(r227983): fast/dom/adopt-node-crash-2.html is flaky https://bugs.webkit.org/show_bug.cgi?id=182589
Reviewed by Wenson Hsieh.

This is a speculative fix for the test since I can't reproduce the flakiness locally. Because r227983 makes
the layout update async, I suspect the focus change which occurs to the keygen having autofocus attribute
isn't updating the layout in time for notifyDone call. Manually update the layout before/after the test.

Also wait for the load event to figure before starting the test since layout, etc... may not be updated
immediately after DOMContentLoaded.

* fast/dom/adopt-node-crash-2.html:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@231690 268f45cc-cd09-0410-ab3c-d52691b4dbfc