Don't use inferred types when the JIT is disabled https://bugs.webkit.org/show_bug.cgi?id=185539
Reviewed by Yusuke Suzuki.

There are many JSC API clients that run with the JIT disabled. They were
all allocating and tracking inferred types for no benefit. Inferred types
only benefit programs when they make it to the DFG/FTL. I was seeing cases
where the inferred type machinery used ~0.5MB. This patch makes is so we
don't allocate that machinery when the JIT is disabled.

* runtime/Structure.cpp:
(JSC::Structure::willStoreValueSlow):
* runtime/Structure.h:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@231703 268f45cc-cd09-0410-ab3c-d52691b4dbfc