Increase the simulated memory size on PLATFORM(IOS_SIMULATOR) from 512MB to 1024MB https://bugs.webkit.org/show_bug.cgi?id=185908
Reviewed by Geoffrey Garen.

Source/bmalloc:

We don't support 512MB devices anymore. This will make the simulator behave more
like a real device.

* bmalloc/AvailableMemory.cpp:
(bmalloc::memorySizeAccordingToKernel):

Factor to a function.
Don't use availableMemoryGuess for the simulator value as it is not a guess.

(bmalloc::computeAvailableMemory):

Apply the same adjustments to the simulated value too.

LayoutTests:

* platform/ios-simulator/fast/canvas/canvas-crash-expected.txt:
* platform/ios-simulator/fast/canvas/canvas-skia-excessive-size-expected.txt:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@232117 268f45cc-cd09-0410-ab3c-d52691b4dbfc