FunctionRareData::m_objectAllocationProfileWatchpoint is racy https://bugs.webkit.org/show_bug.cgi?id=186237
Reviewed by Saam Barati.

We initialize it blind and let it go into auto-watch mode once the DFG adds a watchpoint, but
that means that we never notice that it fired if it fires between when the DFG decides to
watch it and when it actually adds the watchpoint.
        
Most watchpoints are initialized watched for this purpose. This one had a somewhat good
reason for being initialized blind: that's how we knew to ignore changes to the prototype
before the first allocation. However, that functionality also arose out of the fact that the
rare data is created lazily and usually won't exist until the first allocation.
        
The fix here is to make the watchpoint go into watched mode as soon as we initialize the
object allocation profile.
        
It's hard to repro this race, however it started causing spurious test failures for me after
bug 164904.

* runtime/FunctionRareData.cpp:
(JSC::FunctionRareData::FunctionRareData):
(JSC::FunctionRareData::initializeObjectAllocationProfile):



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@232598 268f45cc-cd09-0410-ab3c-d52691b4dbfc