[Cocoa] Web Automation: include browser name and version in listing for automation targets https://bugs.webkit.org/show_bug.cgi?id=186204 <rdar://problem/36950423>
Reviewed by Darin Adler.

Source/JavaScriptCore:

Ask the client what the reported browser name and version should be, then
send this as part of the listing for an automation target.

* inspector/remote/RemoteInspectorConstants.h:
* inspector/remote/cocoa/RemoteInspectorCocoa.mm:
(Inspector::RemoteInspector::listingForAutomationTarget const):

Source/WebKit:

Add a new delegate method that allows the client to set the name and version
of the browser as returned in the 'browserName' and 'browserVersion' capabilities.
If the delegate methods are not implemented, try to get this information from
the main bundle.

In the RWI protocol, these fields are added to automation target listings.

* UIProcess/API/Cocoa/_WKAutomationDelegate.h:
* UIProcess/Cocoa/AutomationClient.h:
* UIProcess/Cocoa/AutomationClient.mm:
(WebKit::AutomationClient::AutomationClient):
(WebKit::AutomationClient::browserName const):
(WebKit::AutomationClient::browserVersion const):

Source/WTF:

* wtf/spi/cf/CFBundleSPI.h: Add needed infoDictionary key values.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@232636 268f45cc-cd09-0410-ab3c-d52691b4dbfc