[JSC] Array.prototype.sort should rejects null comparator https://bugs.webkit.org/show_bug.cgi?id=186458
Reviewed by Keith Miller.

JSTests:

* ChakraCore/test/Array/array_sort.baseline-jsc:
* stress/array-sort-bad-comparator.js:
(test):
* stress/sort-null-comparator.js: Removed.
* test262/expectations.yaml:

Source/JavaScriptCore:

This relaxed behavior is once introduced in r216169 to fix some pages by aligning
the behavior to Chrome and Firefox.

However, now Chrome, Firefox and Edge reject a null comparator. So only JavaScriptCore
accepts it. This patch reverts r216169 to align JSC to the other engines and fix
the spec issue.

* builtins/ArrayPrototype.js:
(sort):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@232666 268f45cc-cd09-0410-ab3c-d52691b4dbfc