Add logging around internalError(const URL&) https://bugs.webkit.org/show_bug.cgi?id=186369 <rdar://problem/40872046>
Reviewed by Brent Fulgham.

There are times when we receive bug reports where the user says that
they are simply shown a page saying an internal error occurred. To
help understand the circumstances of that error, add some logging to
internalError() in WebErrors.cpp. This logging logs at the Error level
that internalError() was called and then logs a backtrace.

Source/WebKit:

* Shared/WebErrors.cpp:
(WebKit::internalError):

Source/WTF:

* wtf/Assertions.cpp:
* wtf/Assertions.h:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@232734 268f45cc-cd09-0410-ab3c-d52691b4dbfc