Test262-Runner: Update config list with some failing tests https://bugs.webkit.org/show_bug.cgi?id=185628
Patch by Leo Balter <leonardo.balter@gmail.com> on 2018-06-14
Reviewed by Michael Saboff.

The config list is updated to match some slow / hanging tests that are currenctly failing
due to the length used in Array or Array like objects.

This also adds matching bugs for known features that are still in progress and not enabled
yet.
* test262/config.yaml:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@232853 268f45cc-cd09-0410-ab3c-d52691b4dbfc