[Win] 'deref': is not a member of 'WebKit::WebResourceLoadStatisticsStore::updateCookiePartitioningForDomains::<lambda_9d761a6dc12d95db7fa2d6f3f5aa26fa>' https://bugs.webkit.org/show_bug.cgi?id=187035
Unreviewed build fix.

MSVC can't compile the code using `this` in a generalized lambda
capture in another lambda.

In this case, there is no need to copy `protectedThis` for the
inner lambda. Move `protectedThis` of the outer lambda to the
inner as well as `completionHandler`.

* UIProcess/WebResourceLoadStatisticsStore.cpp:
(WebKit::WebResourceLoadStatisticsStore::updateCookiePartitioning):
Moved `protectedThis` from the outer lambda to the inner.
(WebKit::WebResourceLoadStatisticsStore::updateCookiePartitioningForDomains):
Ditto.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@233195 268f45cc-cd09-0410-ab3c-d52691b4dbfc