[LFC] Computed height for in-flow non-replaced should not include padding and border. https://bugs.webkit.org/show_bug.cgi?id=187031
Reviewed by Antti Koivisto.

In certain cases the height of a non-replaced in-flow box is computed using the bottom position of its last in-flow child.
The in-flow child's bottom position is in the coordinate system of the containing block's border box (border box's top left is 0, 0) ->
it includes both the (top) border and the padding of the containing block.

* layout/blockformatting/BlockFormattingContextGeometry.cpp:
(WebCore::Layout::BlockFormattingContext::Geometry::inFlowNonReplacedHeightAndMargin):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@233198 268f45cc-cd09-0410-ab3c-d52691b4dbfc